Remove unused length property invocation
This commit removes the unused invocation of 'length' property on the string object in function getSeparator within students.ts file. This redundant invocation took place before the initiation of a for-loop, and it was not used anywhere inside the loop or function. Removing it aids in code readability and optimization.